Rename complaint.rs to recovery_proof.rs
